Create Analysis.py
Uses Scorer/Interpreter to return "graph" of positivity v. time.
The example above takes in three tweets, 7.0, 5.0, and 3.0 days old respectively,
and asks for the predicted positivities for the day before yesterday, yesterday, today, and tomorrow.

This is returned as a list of time-score tuples, e.g. (1.0, -1.5006) for a negative score tomorrow.